+++ Example/AcknowExample/ViewController.swift
@@ -29,7 +29,6 @@ class ViewController: UIViewController {
 
     @IBAction func pushAcknowListWithDefaultStyle(_ sender: AnyObject) {
         let viewController = AcknowListViewController(fileNamed: "Pods-acknowledgements")
-        //viewController.title = "Default Style"
         viewController.headerText = "Visit: https://developer.apple.com"
         navigationController?.pushViewController(viewController, animated: true)
     }