+++ Examples/AcknowExampleCocoaPods/Pods/Target Support Files/AcknowList/AcknowList-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_AcknowList : NSObject
+@end
+@implementation PodsDummy_AcknowList
+@end