+++ Examples/AcknowExampleCocoaPods/Pods/Target Support Files/Pods-AcknowExampleCocoaPods/Pods-AcknowExampleCocoaPods-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_Pods_AcknowExampleCocoaPods : NSObject
+@end
+@implementation PodsDummy_Pods_AcknowExampleCocoaPods
+@end