+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - Open any URL in `AcknowListViewController` header and footer
 - Update `AcknowListViewController` initializers with optional table view style parameter, by Matt Croxson (@Lumus)
-- TO DO: Support SPM localized resources
+- Update supported platforms to iOS 9
+- WIP: Support SPM localized resources
 
 
 ## 1.9.5 (17 September 2020)